Board Layout Template
This is the template I had in mind for the board layout that we should
both work from.

Feel free to change anything you want, it's your board, I just had to
create something to check the grid on.

I'll try to post my progress with the Grid by the end of the weekend.
Thanks!