Implemented final grid layout
I played with the size of the board and buttons a bit and came up with
this layout, I think it looks pretty good. I wanted to make it larger
than the original game because the original was tiny, but also wanted to
keep it a bit slim.

Feel free to modify it as you please though, just let me know if you
change dimensions so I can change my buttons accordingly (try to keep a
difference of multiples of 10).

Original minesweeper board:
181*269
16*16 each square

Thanks!