Icons, Bombs, and MouseEvents
Renamed image files to make more sense. Initialized all icons for the grid, added MouseEvents to each square of the grid, and created the array of random bombs.

For your use, added a public int gameStart, that equals 1 when the game has started (the user left clicks), and 0 when a game is not in progress. Useful for starting the timer.

Also, modified the Frame to be not resizable. It is way too complicated to scale everything.

Since the project is due Tuesday, I'll email you by the weekend to see if you want to get together to work on finishing touches on Monday.
Otherwise, I'm planning to be done with my part by Sunday morning. Will keep you updated!