Updated Board Smileys and Fixes #3 (bug)
I updated the Board's mousePressed to be able to press on all smileys
because we have to be able to reset the game at any time.

Also, I looked into bug you posted and everything looks fine now.
From what I understood, you start up the game, place a flag down, then
left click on an icon to reveal it, and the timer doesn't start.
I tested it with fresh starts, restarts, negative flags, etc, seems to
work in every case now, hopefully it's fixed on your end as well